+++ logicforge/plugins/cats/rule_loop.py
@@ -9,6 +9,9 @@ from logicforge.plugins.cats.adjacent_color_pair_exclusion_rule import (
 from logicforge.plugins.cats.color_confined_to_line_rule import (
     ColorConfinedToLineRule,
 )
+from logicforge.plugins.cats.color_subset_confined_to_lines_rule import (
+    ColorSubsetConfinedToLinesRule,
+)
 from logicforge.plugins.cats.monochromatic_line_color_exclusion_rule import (
     MonochromaticLineColorExclusionRule,
 )
@@ -31,6 +34,7 @@ DEFAULT_CATS_RULES: tuple[CatsApplyRule, ...] = (
     SingleRemainingColorCellRule(),
     SingleRemainingLineCellRule(),
     MonochromaticLineColorExclusionRule(),
+    ColorSubsetConfinedToLinesRule(),
     AdjacentColorPairExclusionRule(),
     ColorConfinedToLineRule(),
 )